chore: fix updating deps on lerna projects
